% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacker_news.r
\name{get_stories}
\alias{get_stories}
\title{This function returns the top stories from hacker news}
\usage{
get_stories(type = "top", limit = 10, pretty_print = FALSE)
}
\arguments{
\item{type}{The type of stories to get, defaults to top stories}

\item{limit}{The top limit stories to return}

\item{pretty_print}{If true return list of serialized stories else return the json structure}
}
\description{
This function returns the top stories from hacker news
}
\examples{
get_stories()
get_stories(type = "top",limit = 10, pretty_print = TRUE)
}

