% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacker_news.r
\name{get_story}
\alias{get_story}
\title{get_story}
\usage{
get_story(id, pretty_print = FALSE)
}
\arguments{
\item{id}{The id of the story to query}

\item{pretty_print}{If true return list of serialized stories else return the json structure}
}
\description{
This function returns the story given it's story id
}
\examples{
get_story(12508356)
get_story(12508356, pretty_print = TRUE)
}

