% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacker_news.r
\name{get_top_stories}
\alias{get_top_stories}
\title{get_top_stories}
\usage{
get_top_stories(limit = 10, pretty_print = FALSE)
}
\arguments{
\item{limit}{The top limit stories to return}

\item{pretty_print}{If true print to the terminal and return NULL else return the json structure}
}
\description{
This function returns the top stories from hacker news
}
\examples{
get_top_stories()
get_top_stories(limit = 5, pretty_print = TRUE)
}

