% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacker_news.r
\name{serialize_story}
\alias{serialize_story}
\title{serialize_story}
\usage{
serialize_story(story)
}
\arguments{
\item{story}{The story to serialize}
}
\description{
This function is responsible for serializing the story to a string
}
