% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacker_news.r
\name{process_stories}
\alias{process_stories}
\title{process_stories}
\usage{
process_stories(stories, pretty_print)
}
\arguments{
\item{stories}{The hacker news stories to process}

\item{pretty_print}{If true return list of serialized stories else return the json structure}
}
\description{
This function processes the stories by either returning the serialized stories or original json structure
}
